init visualization
