+++ src/visualization/assets/style.css
@@ -1,0 +1,4 @@
+/* Add your custom styles here */
+body {
+    font-family: Arial, sans-serif;
+}
